[libc++][pstl] Promote CPU backends to top-level backends
This patch removes the two-level backend dispatching mechanism we
had in the PSTL. Instead of selecting both a PSTL backend and a PSTL
CPU backend, we now only select a top-level PSTL backend. This greatly
simplifies the PSTL configuration layer.

While this patch technically removes some flexibility from the PSTL
configuration mechanism because CPU backends are not considered
separately, it opens the door to a much more powerful configuration
mechanism based on chained backends in a follow-up patch.

This is a step towards overhauling the PSTL dispatching mechanism.
